Commented the code and added some bug fixes.
Previous code was never compiled under -Wall -Wextra, neither was an ASan run tested.
Both have been implemented now and the code passes all test cases - which include a mixture of working and malformed code. Next, we make Value a tagged union.

+++ debug.h
@@ -3,4 +3,6 @@
 #include "chunk.h"
 
 void disassembleChunk(const Chunk& chunk, const char* name);
-int disassembleInstruction(const Chunk& chunk, int offset);
+
+//returns offset of the next instruction since they're variable length. 
+size_t disassembleInstruction(const Chunk& chunk, size_t offset);
